# backend/docs/swagger.yaml
definitions:
...
  github_com_raids-lab_crater_dao_model.GpuAnalysis:
    properties:
      command:
        description: 采集到的原始数据
        type: string
      createdAt:
        description: 自动追踪的时间戳
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      historicalMetrics:
        type: string
      id:
        type: integer
      jobID:
        type: integer
      jobName:
        type: string
      llmversion:
        type: string
      namespace:
        type: string
      phase1LLMReason:
        type: string
      phase1Score:
        description: LLM 分析结果
        type: integer
      phase2LLMReason:
        type: string
      phase2Score:
        type: integer
      podName:
        description: 原始 Kubernetes 信息
        type: string
      reviewStatus:
        allOf:
        - $ref: '#/definitions/github_com_raids-lab_crater_dao_model.ReviewStatus'
        description: 管理状态
      userID:
        type: integer
      userName:
        type: string
    type: object
...
  github_com_raids-lab_crater_dao_model.ReviewStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    format: int32
    type: integer
    x-enum-comments:
      _: 零值，被忽略
      ReviewStatusConfirmed: 已确认
      ReviewStatusIgnored: 已忽略
      ReviewStatusPending: 待审核
    x-enum-descriptions:
    - 零值，被忽略
    - 待审核
    - 已确认
    - 已忽略
    x-enum-varnames:
    - _
    - ReviewStatusPending
    - ReviewStatusConfirmed
    - ReviewStatusIgnored
...
  github_com_raids-lab_crater_internal_resputil.Response-any:
    properties:
      code:
        description: 依然保持 int (ErrorCode) 类型
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  github_com_raids-lab_crater_internal_resputil.Response-github_com_raids-lab_crater_dao_model_GpuAnalysis:
    properties:
      code:
        description: 依然保持 int (ErrorCode) 类型
        type: integer
      data:
        $ref: '#/definitions/github_com_raids-lab_crater_dao_model.GpuAnalysis'
      msg:
        type: string
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  internal_handler.TriggerJobAnalysisReq:
    properties:
      jobName:
        type: string
    required:
    - jobName
    type: object
...
info:
  contact: {}
  description: This is the API server for Crater, a Multi-tenant AI Model Training
    Platform based on Kubernetes.
  title: Crater API
  version: 1.0.0
paths:
...
  /v1/admin/gpu-analysis/trigger/job:
    post:
      consumes:
      - application/json
      description: 提供 Job 的 name 来启动一次完整的分析。如果该 Job 已有分析记录，则更新；否则创建新记录。
      parameters:
      - description: Job 信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_handler.TriggerJobAnalysisReq'
      produces:
      - application/json
      responses:
        "200":
          description: 分析成功，返回创建或更新后的分析记录
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-github_com_raids-lab_crater_dao_model_GpuAnalysis'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: 分析过程中发生错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 手动触发对单个 Job 的 GPU 分析
      tags:
      - GpuAnalysis
...
  /v1/spjobs/{name}/events:
    get:
      consumes:
      - application/json
      description: 获取稀疏推荐作业关联的事件信息
      parameters:
      - description: Job Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 事件列表
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: 其他错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 获取稀疏推荐作业的事件
      tags:
      - SpJob
...
securityDefinitions:
  Bearer:
    description: 访问 /login 并获取 TOKEN 后，填入 'Bearer ${TOKEN}' 以访问受保护的接口
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"